HW3 code samples

+++ code_samples/lec5/fp_precision.cpp
@@ -20,8 +20,8 @@ const unsigned long PRINT_INTERVAL = 1e+06;   // one million
 
 int main(int argc, char *argv[]) {
 
-  float i = 0; 
-  int j = 0;
+  unsigned i = 0; 
+  unsigned j = 0;
   while (i  <= LOOP_BOUND) {
     // print counter value every millionth iteration
     if (j % PRINT_INTERVAL == 0)
